상품페이지 추가

+++ src/main/java/com/pubble/conpub/domain/Mileage.java
@@ -30,7 +30,7 @@ public class Mileage {
     @JoinColumn(name = "member_no")
     private Member mileageMember;
 
-    private BigDecimal mileage;
+    private int mileage;
 
     private String mileagePath;
